Add discard test
