+++ clicktests/test.noheader.js
@@ -44,4 +44,9 @@ suite.test('Check for refresh button', function(done) {
 });
 
 
+suite.test('Shutdown', function(done) {
+  environment.shutdown(done);
+});
+
+
 testsuite.runAllSuits();
